requirements.txt
anyio==3.7.1
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.70.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.22
pyparsing==3.2.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.16.0
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.15.0
